Vitest shared config created in root folder vitest configs extracted from packages modules into one shared config
