+++ package.json
@@ -43,11 +43,11 @@
     "test:main": "vitest run -r packages/main --passWithNoTests --coverage",
     "test:preload": "vitest run -r packages/preload --passWithNoTests --coverage",
     "test:preload-docker-extension": "vitest run -r packages/preload-docker-extension --passWithNoTests --coverage",
-    "test:extensions": "npm run test:extensions:crc && npm run test:extensions:kind && npm run test:extensions:docker && npm run test:extensions:lima && npm run test:extensions:kube && npm run test:extensions:podman && npm run test:extensions:registries",
+    "test:extensions": "npm run test:extensions:compose && npm run test:extensions:kind && npm run test:extensions:docker && npm run test:extensions:lima && npm run test:extensions:kube && npm run test:extensions:podman && npm run test:extensions:registries",
     "test:extensions:kind": "vitest run -r extensions/kind --passWithNoTests --coverage ",
-    "test:extensions:crc": "vitest run -r extensions/crc --passWithNoTests --coverage",
+    "test:extensions:compose": "vitest run -r extensions/compose --passWithNoTests --coverage",
     "test:extensions:docker": "vitest run -r extensions/docker --passWithNoTests --coverage ",
-    "test:extensions:kube": "vitest run -r extensions/kube --passWithNoTests --coverage ",
+    "test:extensions:kube": "vitest run -r extensions/kube-context --passWithNoTests --coverage ",
     "test:extensions:lima": "vitest run -r extensions/lima --passWithNoTests --coverage ",
     "test:extensions:podman": "vitest run -r extensions/podman --passWithNoTests --coverage ",
     "test:extensions:registries": "vitest run -r extensions/registries --passWithNoTests --coverage ",
