+++ vitest-shared-extensions.config.js
